+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/actions/cart/impl/CartImpl.java
@@ -22,7 +22,10 @@ package com.adobe.aem.commons.assetshare.components.actions.cart.impl;
 import com.adobe.aem.commons.assetshare.components.actions.ActionHelper;
 import com.adobe.aem.commons.assetshare.components.actions.cart.Cart;
 import com.adobe.aem.commons.assetshare.content.AssetModel;
+import com.adobe.cq.export.json.ComponentExporter;
+import com.adobe.cq.export.json.ExporterConstants;
 import org.apache.sling.api.SlingHttpServletRequest;
+import org.apache.sling.models.annotations.Exporter;
 import org.apache.sling.models.annotations.Model;
 import org.apache.sling.models.annotations.Required;
 import org.apache.sling.models.annotations.injectorspecific.OSGiService;
@@ -34,9 +37,10 @@ import java.util.Collection;
 
 @Model(
         adaptables = {SlingHttpServletRequest.class},
-        adapters = {Cart.class},
+        adapters = {Cart.class, ComponentExporter.class},
         resourceType = {CartImpl.RESOURCE_TYPE}
 )
+@Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME, extensions = ExporterConstants.SLING_MODEL_EXTENSION)
 public class CartImpl implements Cart {
     protected static final String RESOURCE_TYPE = "asset-share-commons/components/modals/cart";