+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/details/impl/ActionButtonsImpl.java
@@ -21,9 +21,12 @@ package com.adobe.aem.commons.assetshare.components.details.impl;
 
 import com.adobe.aem.commons.assetshare.components.details.ActionButtons;
 import com.adobe.aem.commons.assetshare.configuration.Config;
+import com.adobe.cq.export.json.ComponentExporter;
+import com.adobe.cq.export.json.ExporterConstants;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.models.annotations.DefaultInjectionStrategy;
+import org.apache.sling.models.annotations.Exporter;
 import org.apache.sling.models.annotations.Model;
 import org.apache.sling.models.annotations.Required;
 import org.apache.sling.models.annotations.injectorspecific.Self;
@@ -31,10 +34,11 @@ import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;
 
 @Model(
         adaptables = {SlingHttpServletRequest.class},
-        adapters = {ActionButtons.class},
+        adapters = {ActionButtons.class, ComponentExporter.class},
         resourceType = ActionButtonsImpl.RESOURCE_TYPE,
         defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL
 )
+@Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME, extensions = ExporterConstants.SLING_MODEL_EXTENSION)
 public class ActionButtonsImpl extends AbstractEmptyTextComponent implements ActionButtons {
     protected static final String RESOURCE_TYPE = "asset-share-commons/components/details/action-buttons";