+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/predicates/impl/FreeformTextPredicatePredicateImpl.java
@@ -23,14 +23,13 @@ import com.adobe.aem.commons.assetshare.components.predicates.AbstractPredicate;
 import com.adobe.aem.commons.assetshare.components.predicates.FreeformTextPredicate;
 import com.adobe.aem.commons.assetshare.search.impl.predicateevaluators.PropertyValuesPredicateEvaluator;
 import com.adobe.aem.commons.assetshare.util.PredicateUtil;
+import com.adobe.cq.export.json.ComponentExporter;
+import com.adobe.cq.export.json.ExporterConstants;
 import org.apache.commons.lang.StringUtils;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ValueMap;
-import org.apache.sling.models.annotations.Default;
-import org.apache.sling.models.annotations.DefaultInjectionStrategy;
-import org.apache.sling.models.annotations.Model;
-import org.apache.sling.models.annotations.Required;
+import org.apache.sling.models.annotations.*;
 import org.apache.sling.models.annotations.injectorspecific.Self;
 import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;
 
@@ -44,10 +43,11 @@ import static java.util.Collections.EMPTY_LIST;
 
 @Model(
         adaptables = {SlingHttpServletRequest.class},
-        adapters = {FreeformTextPredicate.class},
+        adapters = {FreeformTextPredicate.class, ComponentExporter.class},
         resourceType = {FreeformTextPredicatePredicateImpl.RESOURCE_TYPE},
         defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL
 )
+@Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME, extensions = ExporterConstants.SLING_MODEL_EXTENSION)
 public class FreeformTextPredicatePredicateImpl extends AbstractPredicate implements FreeformTextPredicate {
     protected static final String RESOURCE_TYPE = "asset-share-commons/components/search/freeform-text";
     private static final String CUSTOM_DELIMITER = "__CUSTOM_DELIMITER";