+++ core/src/main/java/com/adobe/aem/commons/assetshare/components/search/impl/StatisticsImpl.java
@@ -3,8 +3,11 @@ package com.adobe.aem.commons.assetshare.components.search.impl;
 import com.adobe.aem.commons.assetshare.components.search.Statistics;
 import com.adobe.aem.commons.assetshare.search.Search;
 import com.adobe.aem.commons.assetshare.util.ModelCache;
+import com.adobe.cq.export.json.ComponentExporter;
+import com.adobe.cq.export.json.ExporterConstants;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.models.annotations.DefaultInjectionStrategy;
+import org.apache.sling.models.annotations.Exporter;
 import org.apache.sling.models.annotations.Model;
 import org.apache.sling.models.annotations.injectorspecific.Self;
 
@@ -12,10 +15,11 @@ import javax.annotation.PostConstruct;
 
 @Model(
         adaptables = {SlingHttpServletRequest.class},
-        adapters = {Statistics.class},
+        adapters = {Statistics.class, ComponentExporter.class},
         resourceType = {StatisticsImpl.RESOURCE_TYPE},
         defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL
 )
+@Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME, extensions = ExporterConstants.SLING_MODEL_EXTENSION)
 public class StatisticsImpl implements Statistics {
     public static final String RESOURCE_TYPE = "asset-share-commons/components/search/statistics";